refactor(modules): wave 3 splits for domain, persist, pickers, and UI taxonomy
Extract action tree/kind helpers, program catalog resolve/mutate, picker submodules, and control-flow edit fields; move picker categories into sqyre-ui-model while keeping domain wire metadata stable.
