+++ crates/sqyre-app/src/add_action.rs
@@ -17,13 +17,13 @@ use crate::widgets::SaveCancel;
 use eframe::egui::{self, Color32, CornerRadius, Key, Sense, Vec2};
 use sqyre_domain::{
     action_templates, action_type_label, blank_action, Action, ActionId, ActionTemplate,
-    KnownVariableNames, ACTION_PICKER_CATEGORIES,
+    KnownVariableNames,
 };
 use sqyre_hotkeys::{MacroHotkeyBridge, ScreenClickBridge};
 use sqyre_persist::ProgramCatalog;
 use sqyre_persist::UserSettings;
 use sqyre_serialize::{action_from_map, action_to_map};
-use sqyre_ui_model::{action_icon_glyph, action_pastel_color};
+use sqyre_ui_model::{action_icon_glyph, action_pastel_color, action_picker_category, ACTION_PICKER_CATEGORIES};
 use sqyre_validate::validate_action;
 use std::collections::HashMap;
 use web_time::{Duration, Instant};
@@ -226,7 +226,7 @@ impl AddActionPicker {
                                     ui.strong(*category);
                                     ui.add_space(4.0);
                                     for tmpl in
-                                        templates.iter().filter(|t| t.category == *category)
+                                        templates.iter().filter(|t| action_picker_category(t.action_type) == *category)
                                     {
                                         let sample = self
                                             .prototype_for(tmpl.action_type)
