--- crates/sqyre-app/src/pickers.rs
@@ -1,1613 +1,0 @@
-//! Shared entity pickers: item icon grids, point / search-area lists, collection cells,
-//! macros, and Focus Window live-window lists.
-
-use crate::data_editor_preview::{paint_grid_overlay_painter, show_file_hover};
-use crate::icon_cache::IconCache;
-use crate::image_view::{self, ImageViewTransform};
-use crate::paint_ctx::CatalogPaint;
-use crate::preview_tooltip::PreviewKind;
-use eframe::egui::{self, Color32, Pos2, Sense, Vec2};
-use sqyre_capture::WindowInfo;
-use sqyre_domain::{CoordinateRef, PROGRAM_DELIMITER};
-use sqyre_persist::ProgramCatalog;
-use std::sync::mpsc::{self, Receiver, TryRecvError};
-use std::thread;
-
-/// Fixed cell size (thumb + padding; no under-icon label).
-const GRID_CELL: f32 = 64.0;
-const GRID_THUMB: f32 = 52.0;
-const GRID_GAP: f32 = 6.0;
-/// Compact cells for Image Search tip edit (remove badge overlay).
-const EDIT_CELL: f32 = 40.0;
-const EDIT_THUMB: f32 = 36.0;
-const EDIT_GAP: f32 = 4.0;
-const REMOVE_BTN: f32 = 12.0;
-const HEADER_SIZE: f32 = 16.0;
-
-/// In-progress collection cell selection (1-based inclusive).
-#[derive(Debug, Clone)]
-pub struct CollectionCellPick {
-    pub program: String,
-    pub collection: String,
-    pub rows: i32,
-    pub cols: i32,
-    /// Current selection; `None` until the user clicks.
-    pub sel: Option<(i32, i32, i32, i32)>,
-    /// Drag start cell while pointer is down (selection mode).
-    drag_anchor: Option<(i32, i32)>,
-    view: ImageViewTransform,
-}
-
-impl CollectionCellPick {
-    pub fn new(
-        program: impl Into<String>,
-        collection: impl Into<String>,
-        rows: i32,
-        cols: i32,
-    ) -> Self {
-        Self {
-            program: program.into(),
-            collection: collection.into(),
-            rows: rows.max(1),
-            cols: cols.max(1),
-            sel: None,
-            drag_anchor: None,
-            view: ImageViewTransform::default(),
-        }
-    }
-
-    pub fn with_initial_sel(mut self, sel: Option<(i32, i32, i32, i32)>) -> Self {
-        self.sel = sel;
-        self
-    }
-
-    pub fn reset_view(&mut self) {
-        self.view.reset();
-    }
-
-    pub fn to_ref(&self) -> Option<CoordinateRef> {
-        let (r1, c1, r2, c2) = self.sel?;
-        Some(CoordinateRef::collection(
-            &self.program,
-            &self.collection,
-            r1,
-            c1,
-            r2,
-            c2,
-        ))
-    }
-}
-
-/// Which modal picker is open from an action edit tip (or similar).
-#[derive(Debug, Default)]
-pub enum ActivePicker {
-    #[default]
-    None,
-    /// Multi-select item targets (`program~item`).
-    Items { search: String, staged: Vec<String> },
-    /// Point or search-area coordinate picker (`kind` selects catalog + result).
-    Coord {
-        kind: CoordKind,
-        search: String,
-        /// Working value shown/edited in the picker.
-        value: String,
-        /// When set, list is replaced by the collection cell grid.
-        cell_pick: Option<CollectionCellPick>,
-        /// Scroll selected row into view once on open.
-        scroll_to_selection: bool,
-    },
-    Macro {
-        search: String,
-        value: String,
-        scroll_to_selection: bool,
-    },
-    /// Live OS windows for Focus Window (`process_path` + `window_title`).
-    Window {
-        search: String,
-        process_path: String,
-        window_title: String,
-        windows: Vec<WindowInfo>,
-        load_error: Option<String>,
-        scroll_to_selection: bool,
-        /// Background `list_open_windows` result; polled each frame while open.
-        pending: Option<Receiver<Result<Vec<WindowInfo>, String>>>,
-    },
-}
-
-impl Clone for ActivePicker {
-    fn clone(&self) -> Self {
-        match self {
-            ActivePicker::None => ActivePicker::None,
-            ActivePicker::Items { search, staged } => ActivePicker::Items {
-                search: search.clone(),
-                staged: staged.clone(),
-            },
-            ActivePicker::Coord {
-                kind,
-                search,
-                value,
-                cell_pick,
-                scroll_to_selection,
-            } => ActivePicker::Coord {
-                kind: *kind,
-                search: search.clone(),
-                value: value.clone(),
-                cell_pick: cell_pick.clone(),
-                scroll_to_selection: *scroll_to_selection,
-            },
-            ActivePicker::Macro {
-                search,
-                value,
-                scroll_to_selection,
-            } => ActivePicker::Macro {
-                search: search.clone(),
-                value: value.clone(),
-                scroll_to_selection: *scroll_to_selection,
-            },
-            ActivePicker::Window {
-                search,
-                process_path,
-                window_title,
-                windows,
-                load_error,
-                scroll_to_selection,
-                pending: _,
-            } => ActivePicker::Window {
-                search: search.clone(),
-                process_path: process_path.clone(),
-                window_title: window_title.clone(),
-                windows: windows.clone(),
-                load_error: load_error.clone(),
-                scroll_to_selection: *scroll_to_selection,
-                // In-flight loads are not cloned.
-                pending: None,
-            },
-        }
-    }
-}
-
-/// Start a background reload of the live window list (no-op if already loading).
-pub fn refresh_window_picker(picker: &mut ActivePicker) {
-    let ActivePicker::Window {
-        load_error,
-        pending,
-        ..
-    } = picker
-    else {
-        return;
-    };
-    if pending.is_some() {
-        return;
-    }
-    *load_error = None;
-    let (tx, rx) = mpsc::channel();
-    thread::spawn(move || {
-        let _ = tx.send(sqyre_capture::list_open_windows());
-    });
-    *pending = Some(rx);
-}
-
-/// Apply a finished background window-list fetch; request repaint while still loading.
-pub fn poll_window_picker_load(picker: &mut ActivePicker, ctx: &egui::Context) {
-    let ActivePicker::Window {
-        windows,
-        load_error,
-        pending,
-        ..
-    } = picker
-    else {
-        return;
-    };
-    let Some(rx) = pending.as_ref() else {
-        return;
-    };
-    match rx.try_recv() {
-        Ok(Ok(list)) => {
-            *windows = list;
-            *load_error = None;
-            *pending = None;
-        }
-        Ok(Err(e)) => {
-            windows.clear();
-            *load_error = Some(e);
-            *pending = None;
-        }
-        Err(TryRecvError::Empty) => {
-            ctx.request_repaint();
-        }
-        Err(TryRecvError::Disconnected) => {
-            windows.clear();
-            *load_error = Some("window list fetch failed".into());
-            *pending = None;
-        }
-    }
-}
-
-/// Open a Focus Window picker and kick off a background window-list fetch.
-pub fn open_window_picker(process_path: &str, window_title: &str) -> ActivePicker {
-    let mut picker = ActivePicker::Window {
-        search: String::new(),
-        process_path: process_path.to_string(),
-        window_title: window_title.to_string(),
-        windows: Vec::new(),
-        load_error: None,
-        scroll_to_selection: true,
-        pending: None,
-    };
-    refresh_window_picker(&mut picker);
-    picker
-}
-
-impl ActivePicker {
-    pub fn is_open(&self) -> bool {
-        !matches!(self, Self::None)
-    }
-
-    fn cell_pick_mut(&mut self) -> Option<&mut Option<CollectionCellPick>> {
-        match self {
-            Self::Coord { cell_pick, .. } => Some(cell_pick),
-            _ => None,
-        }
-    }
-
-    fn coord_kind(&self) -> Option<CoordKind> {
-        match self {
-            Self::Coord { kind, .. } => Some(*kind),
-            _ => None,
-        }
-    }
-}
-
-/// Fuzzy match on `name` or any tag. Empty `q` matches everything.
-pub fn query_matches_name_or_tags(q: &str, name: &str, tags: &[String]) -> bool {
-    if q.is_empty() {
-        return true;
-    }
-    fuzzy_match_fold(q, name) || tags.iter().any(|t| fuzzy_match_fold(q, t))
-}
-
-/// Subsequence fuzzy match: each needle char appears in order in haystack.
-pub fn fuzzy_match_fold(needle: &str, haystack: &str) -> bool {
-    if needle.is_empty() {
-        return true;
-    }
-    let mut hay = haystack.chars().flat_map(|c| c.to_lowercase());
-    'outer: for nc in needle.chars().flat_map(|c| c.to_lowercase()) {
-        for hc in hay.by_ref() {
-            if hc == nc {
-                continue 'outer;
-            }
-        }
-        return false;
-    }
-    true
-}
-
-fn query_matches_window(q: &str, w: &WindowInfo) -> bool {
-    if q.is_empty() {
-        return true;
-    }
-    fuzzy_match_fold(q, &w.title)
-        || fuzzy_match_fold(q, &w.process_name)
-        || fuzzy_match_fold(q, &w.process_path)
-}
-
-/// Display name + tags for an item target (`program~item` or `program~item~variant`).
-fn item_tooltip_parts(catalog: &ProgramCatalog, target: &str) -> (String, Vec<String>) {
-    let Some((program, rest)) = target.split_once(PROGRAM_DELIMITER) else {
-        return (target.to_string(), Vec::new());
-    };
-    let item_key = rest
-        .split_once(PROGRAM_DELIMITER)
-        .map(|(base, _)| base)
-        .unwrap_or(rest);
-    if let Some(item) = catalog.get(program).and_then(|p| p.items.get(item_key)) {
-        let name = if item.name.is_empty() {
-            item_key.to_string()
-        } else {
-            item.name.clone()
-        };
-        return (name, item.tags.clone());
-    }
-    (item_key.to_string(), Vec::new())
-}
-
-/// Rich hover tooltip: bold name, 12×12 variant icons, then italic primary-colored tags.
-pub fn attach_item_icon_tooltip(
-    response: &egui::Response,
-    catalog: &ProgramCatalog,
-    icons: &mut IconCache,
-    target: &str,
-) {
-    if !response.hovered() {
-        return;
-    }
-    let (name, tags) = item_tooltip_parts(catalog, target);
-    response.clone().on_hover_ui(|ui| {
-        paint_item_icon_tooltip(ui, catalog, icons, target, &name, &tags);
-    });
-}
-
-const VARIANT_TIP_THUMB: f32 = 12.0;
-
-fn paint_item_icon_tooltip(
-    ui: &mut egui::Ui,
-    catalog: &ProgramCatalog,
-    icons: &mut IconCache,
-    target: &str,
-    name: &str,
-    tags: &[String],
-) {
-    ui.set_max_width(280.0);
-    ui.label(egui::RichText::new(name).strong().size(13.0));
-
-    let paths = crate::demo_icons::merged_variant_paths(catalog, target);
-    if !paths.is_empty() {
-        ui.add_space(4.0);
-        ui.horizontal_wrapped(|ui| {
-            ui.spacing_mut().item_spacing = Vec2::splat(2.0);
-            for path in &paths {
-                let Some(tex) = icons.for_path(ui.ctx(), path) else {
-                    continue;
-                };
-                let [tw, th] = tex.size();
-                let size = image_view::fit_in_box(
-                    tw as f32,
-                    th as f32,
-                    VARIANT_TIP_THUMB,
-                    VARIANT_TIP_THUMB,
-                );
-                ui.add(
-                    egui::Image::new((tex.id(), size))
-                        .fit_to_exact_size(size)
-                        .maintain_aspect_ratio(true),
-                );
-            }
-        });
-    }
-
-    if tags.is_empty() {
-        return;
-    }
-    ui.add_space(4.0);
-    let color = ui.visuals().hyperlink_color;
-    for tag in tags {
-        ui.label(egui::RichText::new(tag).size(11.0).italics().color(color));
-    }
-}
-
-fn grid_column_count_for_width(avail_w: f32, cell: f32, gap: f32) -> usize {
-    let avail = avail_w.max(cell);
-    let cols = ((avail + gap) / (cell + gap)).floor() as usize;
-    cols.max(1)
-}
-
-fn icon_grid_metrics(show_remove: bool) -> (f32, f32, f32) {
-    if show_remove {
-        (EDIT_CELL, EDIT_THUMB, EDIT_GAP)
-    } else {
-        (GRID_CELL, GRID_THUMB, GRID_GAP)
-    }
-}
-
-/// Paint a selectable icon cell (fixed square, no under-icon label).
-/// Returns `(cell_clicked, remove_clicked)`.
-pub fn icon_grid_cell_ex(
-    ui: &mut egui::Ui,
-    catalog: &ProgramCatalog,
-    icons: &mut IconCache,
-    target: &str,
-    selected: bool,
-    show_remove: bool,
-) -> (bool, bool) {
-    let (cell, thumb, _) = icon_grid_metrics(show_remove);
-    let rounding = if show_remove { 3.0 } else { 4.0 };
-
-    let mut remove_clicked = false;
-    let desired = Vec2::splat(cell);
-    let (rect, resp) = ui.allocate_exact_size(desired, Sense::click());
-
-    let fill = if selected {
-        Color32::from_rgba_unmultiplied(80, 160, 100, 60)
-    } else if resp.hovered() {
-        Color32::from_black_alpha(25)
-    } else {
-        Color32::TRANSPARENT
-    };
-    let body = rect;
-    ui.painter().rect_filled(body, rounding, fill);
-    if selected {
-        ui.painter().rect_stroke(
-            body,
-            rounding,
-            egui::Stroke::new(2.0, Color32::from_rgb(60, 140, 80)),
-            egui::StrokeKind::Outside,
-        );
-    }
-
-    let tex = icons.for_target_or_fallback(ui.ctx(), catalog, target);
-    let [tw, th] = tex.size();
-    let size = image_view::fit_in_box(tw as f32, th as f32, thumb, thumb);
-    let img_rect = egui::Rect::from_center_size(body.center(), size);
-    // Paint directly — avoid `ui.put(Image)` which can advance the wrap cursor.
-    ui.painter().image(
-        tex.id(),
-        img_rect,
-        egui::Rect::from_min_max(egui::pos2(0.0, 0.0), egui::pos2(1.0, 1.0)),
-        Color32::WHITE,
-    );
-
-    if show_remove {
-        let btn_rect = egui::Rect::from_center_size(
-            egui::pos2(
-                body.right() - REMOVE_BTN * 0.35,
-                body.top() + REMOVE_BTN * 0.35,
-            ),
-            Vec2::splat(REMOVE_BTN),
-        );
-        let btn_id = ui.id().with(("icon_rm", target));
-        let btn_resp = ui.interact(btn_rect, btn_id, Sense::click());
-        let btn_fill = if btn_resp.hovered() {
-            Color32::from_rgb(180, 60, 60)
-        } else {
-            Color32::from_gray(100)
-        };
-        ui.painter()
-            .circle_filled(btn_rect.center(), REMOVE_BTN * 0.5, btn_fill);
-        crate::theme::paint_text_centered(
-            ui,
-            btn_rect,
-            "×",
-            egui::FontId::proportional(9.0),
-            Color32::WHITE,
-        );
-        remove_clicked = btn_resp.clicked();
-    }
-
-    attach_item_icon_tooltip(&resp, catalog, icons, target);
-
-    (resp.clicked() && !remove_clicked, remove_clicked)
-}
-
-/// Lay out `targets` in fixed-size rows (no column stretch, no staircase wrap).
-#[allow(clippy::too_many_arguments)]
-pub fn paint_even_icon_grid(
-    ui: &mut egui::Ui,
-    catalog: &ProgramCatalog,
-    icons: &mut IconCache,
-    targets: &[String],
-    is_selected: impl Fn(&str) -> bool,
-    show_remove: bool,
-    mut on_cell: impl FnMut(usize, &str),
-    mut on_remove: impl FnMut(usize),
-) {
-    if targets.is_empty() {
-        return;
-    }
-    let (cell, _, gap) = icon_grid_metrics(show_remove);
-    let avail = ui.available_width().max(cell);
-    ui.set_max_width(avail);
-    let cols = grid_column_count_for_width(avail, cell, gap);
-    let old_spacing = ui.spacing().item_spacing;
-    ui.spacing_mut().item_spacing = Vec2::splat(gap);
-
-    let mut i = 0;
-    while i < targets.len() {
-        ui.allocate_ui_with_layout(
-            egui::vec2(avail, cell),
-            egui::Layout::left_to_right(egui::Align::Center),
-            |ui| {
-                ui.set_max_width(avail);
-                ui.spacing_mut().item_spacing = Vec2::splat(gap);
-                let end = (i + cols).min(targets.len());
-                for (k, target) in targets.iter().enumerate().take(end).skip(i) {
-                    let sel = is_selected(target);
-                    let (clicked, remove) =
-                        icon_grid_cell_ex(ui, catalog, icons, target, sel, show_remove);
-                    if clicked {
-                        on_cell(k, target);
-                    }
-                    if remove {
-                        on_remove(k);
-                    }
-                }
-            },
-        );
-        i += cols;
-    }
-
-    ui.spacing_mut().item_spacing = old_spacing;
-}
-
-/// Persistent collapse id for a program group in [`paint_items_icon_grid`].
-pub fn items_icon_grid_collapse_id(program: &str) -> egui::Id {
-    egui::Id::new(("items_icon_grid", program))
-}
-
-/// Set open/closed for every program group in the items icon grid.
-pub fn set_items_icon_grid_openness<'a>(
-    ctx: &egui::Context,
-    programs: impl IntoIterator<Item = &'a str>,
-    open: bool,
-) {
-    set_collapsing_openness(
-        ctx,
-        programs.into_iter().map(items_icon_grid_collapse_id),
-        open,
-    );
-}
-
-/// Store open/closed for each collapsing id (expand/collapse-all).
-pub fn set_collapsing_openness(
-    ctx: &egui::Context,
-    ids: impl IntoIterator<Item = egui::Id>,
-    open: bool,
-) {
-    for id in ids {
-        let mut state =
-            egui::collapsing_header::CollapsingState::load_with_default_open(ctx, id, true);
-        state.set_open(open);
-        state.store(ctx);
-    }
-}
-
-/// Program accordion of item icon grids. Click toggles membership in `selected` when
-/// `multi` is true; otherwise replaces selection with the clicked target.
-/// When `multi`, each program header includes an All control over filtered targets
-/// (items picker tri-state / All button).
-///
-/// When `selected_program` / `clicked_program` are used (data editor), program headers
-/// are selectable and write the clicked program name into `clicked_program`.
-#[allow(clippy::too_many_arguments)]
-pub fn paint_items_icon_grid(
-    ui: &mut egui::Ui,
-    catalog: &ProgramCatalog,
-    icons: &mut IconCache,
-    search: &str,
-    selected: &mut Vec<String>,
-    multi: bool,
-    selected_program: Option<&str>,
-    clicked_program: &mut Option<String>,
-) {
-    let q = search.trim().to_ascii_lowercase();
-    let pane_w = ui.available_width();
-    ui.set_max_width(pane_w);
-    for prog in catalog.program_names() {
-        let Some(pdata) = catalog.get(prog) else {
-            continue;
-        };
-        let items: Vec<_> = pdata
-            .items
-            .iter()
-            .filter(|(name, item)| {
-                if q.is_empty() {
-                    return true;
-                }
-                fuzzy_match_fold(&q, prog) || query_matches_name_or_tags(&q, name, &item.tags)
-            })
-            .map(|(name, _)| name.clone())
-            .collect();
-        if items.is_empty() {
-            continue;
-        }
-        let mut targets: Vec<(String, String)> = items
-            .iter()
-            .map(|item_key| {
-                let display = pdata
-                    .items
-                    .get(item_key)
-                    .map(|it| {
-                        if it.name.trim().is_empty() {
-                            item_key.clone()
-                        } else {
-                            it.name.clone()
-                        }
-                    })
-                    .unwrap_or_else(|| item_key.clone());
-                (format!("{prog}{PROGRAM_DELIMITER}{item_key}"), display)
-            })
-            .collect();
-        sort_by_display_name(&mut targets);
-        let targets: Vec<String> = targets.into_iter().map(|(t, _)| t).collect();
-
-        let selected_in_group = targets
-            .iter()
-            .filter(|t| selected.iter().any(|s| s == *t))
-            .count();
-        let all_label = if selected_in_group == 0 {
-            "All"
-        } else if selected_in_group == targets.len() {
-            "None"
-        } else {
-            "All"
-        };
-
-        // Absolute id so expand/collapse-all (outside this ui stack) can target the same state.
-        let id = items_icon_grid_collapse_id(prog);
-        egui::collapsing_header::CollapsingState::load_with_default_open(ui.ctx(), id, true)
-            .show_header(ui, |ui| {
-                let prog_selected = selected_program == Some(prog.as_str());
-                if crate::icon_cache::paint_program_label(
-                    ui,
-                    catalog,
-                    icons,
-                    prog,
-                    crate::icon_cache::ProgramLabelStyle::Header {
-                        selected: Some(prog_selected),
-                        child_count: pdata.items.len(),
-                    },
-                )
-                .clicked()
-                {
-                    *clicked_program = Some(prog.clone());
-                }
-            })
-            .body(|ui| {
-                ui.set_max_width(pane_w);
-                if multi {
-                    ui.horizontal(|ui| {
-                        if ui
-                            .small_button(all_label)
-                            .on_hover_text(
-                                "Select all visible items, or deselect all if all visible are selected",
-                            )
-                            .clicked()
-                        {
-                            toggle_select_all_filtered(selected, &targets);
-                        }
-                        if selected_in_group > 0 && selected_in_group < targets.len() {
-                            ui.weak(format!("{selected_in_group}/{}", targets.len()));
-                        }
-                    });
-                }
-                let mut clicked: Option<String> = None;
-                paint_even_icon_grid(
-                    ui,
-                    catalog,
-                    icons,
-                    &targets,
-                    |t| selected.iter().any(|s| s == t),
-                    false,
-                    |_i, t| {
-                        clicked = Some(t.to_string());
-                    },
-                    |_| {},
-                );
-                if let Some(target) = clicked {
-                    let is_sel = selected.iter().any(|t| t == &target);
-                    if multi {
-                        if is_sel {
-                            selected.retain(|t| t != &target);
-                        } else {
-                            selected.push(target);
-                        }
-                    } else {
-                        *selected = vec![target];
-                    }
-                }
-            });
-    }
-}
-
-/// Toggle: if every `filtered` target is selected, remove them; otherwise add missing ones.
-fn toggle_select_all_filtered(selected: &mut Vec<String>, filtered: &[String]) {
-    if filtered.is_empty() {
-        return;
-    }
-    let all_selected = filtered.iter().all(|t| selected.iter().any(|s| s == t));
-    if all_selected {
-        selected.retain(|s| !filtered.iter().any(|t| t == s));
-    } else {
-        for t in filtered {
-            if !selected.iter().any(|s| s == t) {
-                selected.push(t.clone());
-            }
-        }
-    }
-}
-
-/// Sort `(key, display_name)` rows by display name (case-insensitive), then key.
-fn sort_by_display_name(rows: &mut [(String, String)]) {
-    rows.sort_by(|a, b| {
-        a.1.to_ascii_lowercase()
-            .cmp(&b.1.to_ascii_lowercase())
-            .then_with(|| a.0.cmp(&b.0))
-    });
-}
-
-fn maybe_scroll_to(ui: &mut egui::Ui, resp: &egui::Response, scroll: &mut bool) {
-    if *scroll {
-        ui.scroll_to_rect(resp.rect, Some(egui::Align::Center));
-        *scroll = false;
-    }
-}
-
-/// Options for [`picker_searchable_scroll`].
-pub struct PickerScrollOpts<'a> {
-    /// Space still laid out below the scroll (Save/Cancel). Use `0` when the list fills the pane.
-    pub footer_reserve: f32,
-    /// Extra widgets after the search field (e.g. Refresh).
-    pub trailing: Option<&'a mut dyn FnMut(&mut egui::Ui)>,
-    pub id_salt: Option<&'static str>,
-}
-
-impl PickerScrollOpts<'_> {
-    /// Popup list with Save/Cancel (or similar) below the scroll.
-    pub fn list(_ui: &egui::Ui) -> Self {
-        Self {
-            footer_reserve: 52.0,
-            trailing: None,
-            id_salt: None,
-        }
-    }
-
-    /// Fixed pane that ends at the list (e.g. data editor left column).
-    pub fn pane() -> Self {
-        Self {
-            footer_reserve: 0.0,
-            trailing: None,
-            id_salt: None,
-        }
-    }
-}
-
-/// Search row → separator → capped vertical scroll. `body` receives lowercase trimmed query.
-///
-/// Scroll height is measured after the search row so the pane fits remaining space.
-/// Returns whether search text changed this frame (callers can re-arm scroll-to-selection).
-pub fn picker_searchable_scroll(
-    ui: &mut egui::Ui,
-    search: &mut String,
-    mut opts: PickerScrollOpts<'_>,
-    mut body: impl FnMut(&mut egui::Ui, &str),
-) -> bool {
-    let mut search_changed = false;
-    ui.horizontal(|ui| {
-        ui.label(egui::RichText::new(egui_phosphor::regular::MAGNIFYING_GLASS).size(HEADER_SIZE))
-            .on_hover_text("Search");
-        if ui.text_edit_singleline(search).changed() {
-            search_changed = true;
-        }
-        if let Some(trailing) = opts.trailing.as_mut() {
-            trailing(ui);
-        }
-    });
-    ui.separator();
-    let q = search.trim().to_ascii_lowercase();
-    // Fixed panes (footer_reserve == 0) use remaining height only — no popup screen cap.
-    let max_h = if opts.footer_reserve <= 0.0 {
-        ui.available_height().max(40.0)
-    } else {
-        popup_scroll_max_height(ui, opts.footer_reserve)
-    };
-    let mut scroll = egui::ScrollArea::vertical().auto_shrink([false, false]);
-    if let Some(salt) = opts.id_salt {
-        scroll = scroll.id_salt(salt);
-    }
-    scroll.max_height(max_h).show(ui, |ui| {
-        body(ui, &q);
-    });
-    search_changed
-}
-
-/// Finite height for scroll panes inside content-sized popup windows.
-/// Without this, `ScrollArea` + `auto_shrink([false, false])` grows the window forever.
-///
-/// `footer_reserve` is space still to be laid out below the scroll (buttons, status).
-pub fn popup_scroll_max_height(ui: &egui::Ui, footer_reserve: f32) -> f32 {
-    const FALLBACK: f32 = 360.0;
-    let screen_cap = (ui.ctx().content_rect().height() * 0.65).max(100.0);
-    let h = ui.available_height() - footer_reserve;
-    let capped = if h.is_finite() { h.max(40.0) } else { FALLBACK };
-    capped.min(screen_cap)
-}
-
-/// Flat searchable list of `program~name` refs from points or search areas,
-/// plus program collections (opens cell picker on click).
-/// Rows are sorted by display name within each program.
-pub fn paint_coord_ref_list(
-    ui: &mut egui::Ui,
-    paint: &mut CatalogPaint<'_>,
-    search: &str,
-    current: &mut String,
-    kind: CoordKind,
-    cell_pick: &mut Option<CollectionCellPick>,
-    scroll_to_selection: &mut bool,
-) {
-    let CatalogPaint {
-        catalog,
-        icons,
-        previews,
-    } = paint;
-    let q = search.trim().to_ascii_lowercase();
-    let res = catalog.resolution_key().to_string();
-    let preview_kind = match kind {
-        CoordKind::Point => PreviewKind::Point,
-        CoordKind::SearchArea => PreviewKind::SearchArea,
-    };
-    let current_ref = CoordinateRef(current.clone());
-    let mut did_scroll = false;
-    let list_h = popup_scroll_max_height(ui, 52.0);
-    egui::ScrollArea::vertical()
-        .auto_shrink([false, false])
-        .max_height(list_h)
-        .show(ui, |ui| {
-            for prog in catalog.program_names() {
-                let Some(pdata) = catalog.get(prog) else {
-                    continue;
-                };
-                #[derive(Clone)]
-                enum Row {
-                    Coord { key: String, display: String },
-                    Collection(sqyre_persist::ProgramCollection),
-                }
-                let mut rows: Vec<(String, Row)> = Vec::new();
-                match kind {
-                    CoordKind::Point => {
-                        if let Some(m) = pdata
-                            .points
-                            .get(&res)
-                            .or_else(|| pdata.points.values().next())
-                        {
-                            for (key, pt) in m {
-                                let display = if pt.name.trim().is_empty() {
-                                    key.clone()
-                                } else {
-                                    pt.name.clone()
-                                };
-                                if q.is_empty()
-                                    || fuzzy_match_fold(&q, key)
-                                    || fuzzy_match_fold(&q, &display)
-                                    || fuzzy_match_fold(&q, prog)
-                                {
-                                    rows.push((
-                                        display.clone(),
-                                        Row::Coord {
-                                            key: key.clone(),
-                                            display,
-                                        },
-                                    ));
-                                }
-                            }
-                        }
-                    }
-                    CoordKind::SearchArea => {
-                        if let Some(m) = pdata
-                            .search_areas
-                            .get(&res)
-                            .or_else(|| pdata.search_areas.values().next())
-                        {
-                            for (key, sa) in m {
-                                let display = if sa.name.trim().is_empty() {
-                                    key.clone()
-                                } else {
-                                    sa.name.clone()
-                                };
-                                if q.is_empty()
-                                    || fuzzy_match_fold(&q, key)
-                                    || fuzzy_match_fold(&q, &display)
-                                    || fuzzy_match_fold(&q, prog)
-                                {
-                                    rows.push((
-                                        display.clone(),
-                                        Row::Coord {
-                                            key: key.clone(),
-                                            display,
-                                        },
-                                    ));
-                                }
-                            }
-                        }
-                    }
-                }
-                for col in pdata.collections.values() {
-                    if q.is_empty() || fuzzy_match_fold(&q, &col.name) || fuzzy_match_fold(&q, prog)
-                    {
-                        rows.push((col.name.clone(), Row::Collection(col.clone())));
-                    }
-                }
-                if rows.is_empty() {
-                    continue;
-                }
-                rows.sort_by(|a, b| {
-                    a.0.to_ascii_lowercase()
-                        .cmp(&b.0.to_ascii_lowercase())
-                        .then_with(|| match (&a.1, &b.1) {
-                            (Row::Coord { key: ka, .. }, Row::Coord { key: kb, .. }) => ka.cmp(kb),
-                            (Row::Collection(ca), Row::Collection(cb)) => ca.name.cmp(&cb.name),
-                            (Row::Coord { .. }, Row::Collection(_)) => std::cmp::Ordering::Less,
-                            (Row::Collection(_), Row::Coord { .. }) => std::cmp::Ordering::Greater,
-                        })
-                });
-
-                crate::icon_cache::paint_program_label(
-                    ui,
-                    catalog,
-                    icons,
-                    prog,
-                    crate::icon_cache::ProgramLabelStyle::Header {
-                        selected: None,
-                        child_count: rows.len(),
-                    },
-                );
-                for (_, row) in rows {
-                    match row {
-                        Row::Coord { key, display } => {
-                            let target = format!("{prog}{PROGRAM_DELIMITER}{key}");
-                            let selected = current == &target;
-                            let label = if display == key {
-                                format!("  {key}")
-                            } else {
-                                format!("  {display}")
-                            };
-                            let resp = ui
-                                .selectable_label(selected, egui::RichText::new(label).size(13.0));
-                            previews.show_for_entity(ui, &resp, catalog, prog, &key, preview_kind);
-                            if selected && *scroll_to_selection && !did_scroll {
-                                maybe_scroll_to(ui, &resp, scroll_to_selection);
-                                did_scroll = true;
-                            }
-                            if resp.clicked() {
-                                *current = target;
-                            }
-                        }
-                        Row::Collection(col) => {
-                            let selected = current_ref.is_collection()
-                                && current_ref.program() == Some(prog.as_str())
-                                && current_ref.name() == col.name;
-                            let label = format!("  {} (collection)", col.name);
-                            let resp = ui
-                                .selectable_label(selected, egui::RichText::new(label).size(13.0));
-                            show_file_hover(
-                                ui,
-                                &resp,
-                                icons,
-                                &catalog.collection_image_path(prog, &col.name),
-                                &format!("{prog}~{}", col.name),
-                            );
-                            if selected && *scroll_to_selection && !did_scroll {
-                                maybe_scroll_to(ui, &resp, scroll_to_selection);
-                                did_scroll = true;
-                            }
-                            if resp.clicked() {
-                                let initial = if selected {
-                                    current_ref.cell_range()
-                                } else {
-                                    None
-                                };
-                                *cell_pick = Some(
-                                    CollectionCellPick::new(prog, &col.name, col.rows, col.cols)
-                                        .with_initial_sel(initial),
-                                );
-                            }
-                        }
-                    }
-                }
-                ui.add_space(6.0);
-            }
-        });
-    if *scroll_to_selection && !did_scroll {
-        // Selection not visible under current filter — don't keep retrying every frame.
-        *scroll_to_selection = false;
-    }
-}
-
-fn fit_panel(w: f32, h: f32) -> Vec2 {
-    crate::data_editor_preview::fit_panel(w, h)
-}
-
-/// Interactive collection image + rows×cols overlay; click/drag selects cells.
-/// Wheel zooms at cursor; when zoomed, drag pans.
-fn paint_collection_cell_picker(
-    ui: &mut egui::Ui,
-    catalog: &ProgramCatalog,
-    icons: &mut IconCache,
-    pick: &mut CollectionCellPick,
-) {
-    ui.horizontal(|ui| {
-        crate::icon_cache::paint_program_icon(ui, catalog, icons, &pick.program);
-        ui.label(
-            egui::RichText::new(format!(
-                "Select cells — {}~{}",
-                pick.program, pick.collection
-            ))
-            .strong(),
-        );
-        ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
-            if ui
-                .add_enabled(
-                    pick.view.needs_reset_button(),
-                    egui::Button::new("Reset view"),
-                )
-                .on_hover_text("Fit image in viewport")
-                .clicked()
-            {
-                pick.reset_view();
-            }
-            if pick.view.shows_zoom_label() {
-                ui.weak(format!("{:.0}%", pick.view.zoom * 100.0));
-            }
-        });
-    });
-    ui.weak("Scroll to zoom; drag to pan when zoomed; click/drag selects cells at 100%.");
-
-    let path = catalog.collection_image_path(&pick.program, &pick.collection);
-    let tex = icons.for_path(ui.ctx(), &path);
-    let avail = ui.available_width().min(520.0);
-    let image_size = match &tex {
-        Some(t) => {
-            let [tw, th] = t.size();
-            Vec2::new(tw as f32, th as f32)
-        }
-        None => Vec2::new(avail, avail * 0.75),
-    };
-    let fit = fit_panel(image_size.x, image_size.y);
-    let scale = (avail / fit.x).min(1.0);
-    let desired = Vec2::new((fit.x * scale).max(160.0), (fit.y * scale).max(120.0));
-    let (viewport, resp) = ui.allocate_exact_size(desired, Sense::click_and_drag());
-
-    image_view::handle_scroll_zoom(ui, viewport, image_size, &mut pick.view, resp.hovered());
-
-    let content =
-        image_view::image_content_rect(viewport, image_size, pick.view.zoom, pick.view.pan);
-
-    {
-        let painter = ui.painter_at(viewport);
-        if let Some(tex) = &tex {
-            painter.image(
-                tex.id(),
-                content,
-                egui::Rect::from_min_max(egui::pos2(0.0, 0.0), egui::pos2(1.0, 1.0)),
-                Color32::WHITE,
-            );
-        } else {
-            painter.rect_filled(viewport, 0.0, Color32::from_gray(40));
-            painter.text(
-                viewport.center(),
-                egui::Align2::CENTER_CENTER,
-                "No collection image",
-                egui::FontId::proportional(14.0),
-                Color32::LIGHT_GRAY,
-            );
-        }
-        paint_grid_overlay_painter(&painter, content, pick.rows, pick.cols);
-        if let Some(sel) = pick.sel {
-            paint_cell_selection_painter(&painter, content, pick.rows, pick.cols, sel);
-        }
-    }
-
-    if image_view::handle_pan_drag(&resp, viewport, image_size, &mut pick.view) {
-        if resp.drag_started() {
-            pick.drag_anchor = None;
-        }
-    } else if let Some(pos) = resp.interact_pointer_pos() {
-        if let Some((r, c)) = cell_at(content, pick.rows, pick.cols, pos) {
-            if resp.drag_started() || (resp.clicked() && !resp.dragged()) {
-                pick.drag_anchor = Some((r, c));
-                pick.sel = Some((r, c, r, c));
-            } else if resp.dragged() {
-                if let Some((ar, ac)) = pick.drag_anchor {
-                    pick.sel = Some((ar, ac, r, c));
-                }
-            }
-        }
-        if resp.drag_stopped() {
-            pick.drag_anchor = None;
-            if let Some((r1, c1, r2, c2)) = pick.sel {
-                let (r1, r2) = if r1 <= r2 { (r1, r2) } else { (r2, r1) };
-                let (c1, c2) = if c1 <= c2 { (c1, c2) } else { (c2, c1) };
-                pick.sel = Some((r1, c1, r2, c2));
-            }
-        }
-    }
-
-    let status = match pick.sel {
-        Some((r1, c1, r2, c2)) if r1 == r2 && c1 == c2 => {
-            format!("Selected cell R{r1} C{c1}")
-        }
-        Some((r1, c1, r2, c2)) => format!("Selected R{r1}–{r2} × C{c1}–{c2}"),
-        None => "Click or drag to select cell(s)".into(),
-    };
-    ui.weak(status);
-}
-
-fn paint_cell_selection_painter(
-    painter: &egui::Painter,
-    rect: egui::Rect,
-    rows: i32,
-    cols: i32,
-    sel: (i32, i32, i32, i32),
-) {
-    let (r1, c1, r2, c2) = sel;
-    let (r1, r2) = if r1 <= r2 { (r1, r2) } else { (r2, r1) };
-    let (c1, c2) = if c1 <= c2 { (c1, c2) } else { (c2, c1) };
-    let rows = rows.max(1) as f32;
-    let cols = cols.max(1) as f32;
-    let cw = rect.width() / cols;
-    let ch = rect.height() / rows;
-    let sel_rect = egui::Rect::from_min_max(
-        egui::pos2(
-            rect.left() + (c1 as f32 - 1.0) * cw,
-            rect.top() + (r1 as f32 - 1.0) * ch,
-        ),
-        egui::pos2(rect.left() + c2 as f32 * cw, rect.top() + r2 as f32 * ch),
-    );
-    painter.rect_filled(
-        sel_rect,
-        0.0,
-        Color32::from_rgba_unmultiplied(60, 160, 255, 70),
-    );
-    painter.rect_stroke(
-        sel_rect,
-        0.0,
-        egui::Stroke::new(2.0, Color32::from_rgb(40, 140, 255)),
-        egui::StrokeKind::Outside,
-    );
-}
-
-fn cell_at(rect: egui::Rect, rows: i32, cols: i32, pos: Pos2) -> Option<(i32, i32)> {
-    if !rect.contains(pos) {
-        return None;
-    }
-    let rows = rows.max(1) as f32;
-    let cols = cols.max(1) as f32;
-    let c = (((pos.x - rect.left()) / rect.width()) * cols).floor() as i32 + 1;
-    let r = (((pos.y - rect.top()) / rect.height()) * rows).floor() as i32 + 1;
-    Some((r.clamp(1, rows as i32), c.clamp(1, cols as i32)))
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum CoordKind {
-    Point,
-    SearchArea,
-}
-
-/// Draw open picker window; returns committed value on Save (and clears picker).
-pub fn show_active_picker(
-    ctx: &egui::Context,
-    picker: &mut ActivePicker,
-    paint: &mut CatalogPaint<'_>,
-    // `(name, tags)` — tags are used by the macro search bar.
-    macros: &[(String, Vec<String>)],
-) -> PickerResult {
-    let mut result = PickerResult::None;
-    let mut open = picker.is_open();
-    if !open {
-        return result;
-    }
-
-    poll_window_picker_load(picker, ctx);
-
-    let in_cell_pick = matches!(
-        picker,
-        ActivePicker::Coord {
-            cell_pick: Some(_),
-            ..
-        }
-    );
-
-    let title = match picker {
-        ActivePicker::Items { .. } => "Pick items",
-        ActivePicker::Coord {
-            cell_pick: Some(_), ..
-        } => "Select collection cells",
-        ActivePicker::Coord {
-            kind: CoordKind::Point,
-            ..
-        } => "Pick point",
-        ActivePicker::Coord {
-            kind: CoordKind::SearchArea,
-            ..
-        } => "Pick search area",
-        ActivePicker::Macro { .. } => "Pick macro",
-        ActivePicker::Window { .. } => "Pick window",
-        ActivePicker::None => return result,
-    };
-
-    let mut save = false;
-    let mut cancel = false;
-    let mut back = false;
-
-    egui::Window::new(title)
-        .collapsible(false)
-        .resizable(true)
-        .default_size([560.0, 460.0])
-        .min_size([400.0, 280.0])
-        .order(egui::Order::Foreground)
-        .open(&mut open)
-        .show(ctx, |ui| {
-            match picker {
-                ActivePicker::Items { search, staged } => {
-                    let mut header_click = None;
-                    picker_searchable_scroll(ui, search, PickerScrollOpts::list(ui), |ui, q| {
-                        paint_items_icon_grid(
-                            ui,
-                            paint.catalog,
-                            paint.icons,
-                            q,
-                            staged,
-                            true,
-                            None,
-                            &mut header_click,
-                        );
-                    });
-                    ui.separator();
-                    ui.label(format!("{} selected", staged.len()));
-                }
-                ActivePicker::Coord {
-                    kind,
-                    search,
-                    value,
-                    cell_pick,
-                    scroll_to_selection,
-                } => {
-                    if let Some(pick) = cell_pick.as_mut() {
-                        paint_collection_cell_picker(ui, paint.catalog, paint.icons, pick);
-                    } else {
-                        // Search chrome only — list owns its own ScrollArea (program groups).
-                        ui.horizontal(|ui| {
-                            ui.label(
-                                egui::RichText::new(egui_phosphor::regular::MAGNIFYING_GLASS)
-                                    .size(HEADER_SIZE),
-                            )
-                            .on_hover_text("Search");
-                            if ui.text_edit_singleline(search).changed() {
-                                *scroll_to_selection = true;
-                            }
-                        });
-                        ui.separator();
-                        paint_coord_ref_list(
-                            ui,
-                            paint,
-                            search,
-                            value,
-                            *kind,
-                            cell_pick,
-                            scroll_to_selection,
-                        );
-                    }
-                }
-                ActivePicker::Macro {
-                    search,
-                    value,
-                    scroll_to_selection,
-                } => {
-                    let mut did_scroll = false;
-                    let search_changed = picker_searchable_scroll(
-                        ui,
-                        search,
-                        PickerScrollOpts::list(ui),
-                        |ui, q| {
-                            for (name, tags) in macros {
-                                if !query_matches_name_or_tags(q, name, tags) {
-                                    continue;
-                                }
-                                let selected = value == name;
-                                let resp = ui.selectable_label(
-                                    selected,
-                                    egui::RichText::new(name.as_str()).size(13.0),
-                                );
-                                if selected && *scroll_to_selection && !did_scroll {
-                                    maybe_scroll_to(ui, &resp, scroll_to_selection);
-                                    did_scroll = true;
-                                }
-                                if resp.clicked() {
-                                    *value = name.clone();
-                                }
-                            }
-                        },
-                    );
-                    if search_changed {
-                        *scroll_to_selection = true;
-                    } else if *scroll_to_selection && !did_scroll {
-                        *scroll_to_selection = false;
-                    }
-                }
-                ActivePicker::Window {
-                    search,
-                    process_path,
-                    window_title,
-                    windows,
-                    load_error,
-                    scroll_to_selection,
-                    pending,
-                } => {
-                    let loading = pending.is_some();
-                    let mut did_scroll = false;
-                    let mut refresh_clicked = false;
-                    let mut opts = PickerScrollOpts::list(ui);
-                    let mut trailing = |ui: &mut egui::Ui| {
-                        refresh_clicked = ui
-                            .add_enabled_ui(!loading, |ui| crate::theme::icon_button(ui, "↻"))
-                            .inner
-                            .on_hover_text(if loading { "Refreshing…" } else { "Refresh" })
-                            .clicked();
-                    };
-                    opts.trailing = Some(&mut trailing);
-                    let search_changed = picker_searchable_scroll(ui, search, opts, |ui, q| {
-                        if loading {
-                            ui.label("Loading windows…");
-                        }
-                        if let Some(err) = load_error.as_ref() {
-                            ui.colored_label(crate::theme::error_fg(), err.as_str());
-                        }
-                        for w in windows.iter() {
-                            if !query_matches_window(q, w) {
-                                continue;
-                            }
-                            let selected =
-                                window_title == &w.title && process_path == &w.process_path;
-                            // Prefer icon bytes from the list fetch; avoid per-row OS re-scan.
-                            let process_tex = match w.icon.as_ref() {
-                                Some(icon) => {
-                                    paint
-                                        .icons
-                                        .seed_process_icon(ui.ctx(), &w.process_path, icon)
-                                }
-                                None => paint.icons.cached_process(&w.process_path),
-                            };
-                            let resp = ui
-                                .horizontal(|ui| {
-                                    if let Some(tex) = process_tex.as_ref() {
-                                        crate::icon_cache::paint_process_icon(
-                                            ui,
-                                            tex,
-                                            crate::icon_cache::PROCESS_ICON_SIDE,
-                                        );
-                                    }
-                                    ui.selectable_label(
-                                        selected,
-                                        egui::RichText::new(w.label()).size(13.0),
-                                    )
-                                })
-                                .inner;
-                            if selected && *scroll_to_selection && !did_scroll {
-                                maybe_scroll_to(ui, &resp, scroll_to_selection);
-                                did_scroll = true;
-                            }
-                            if resp.clicked() {
-                                *window_title = w.title.clone();
-                                *process_path = w.process_path.clone();
-                            }
-                        }
-                    });
-                    if refresh_clicked && pending.is_none() {
-                        *load_error = None;
-                        let (tx, rx) = mpsc::channel();
-                        thread::spawn(move || {
-                            let _ = tx.send(sqyre_capture::list_open_windows());
-                        });
-                        *pending = Some(rx);
-                        *scroll_to_selection = true;
-                        ui.ctx().request_repaint();
-                    } else if search_changed {
-                        *scroll_to_selection = true;
-                    } else if *scroll_to_selection && !did_scroll {
-                        *scroll_to_selection = false;
-                    }
-                }
-                ActivePicker::None => {}
-            }
-
-            ui.separator();
-            let cell_has_sel = picker
-                .cell_pick_mut()
-                .and_then(|c| c.as_ref())
-                .and_then(|p| p.sel)
-                .is_some();
-            ui.horizontal(|ui| {
-                if in_cell_pick && ui.button("Back").clicked() {
-                    back = true;
-                }
-                if ui.button("Cancel").clicked() {
-                    cancel = true;
-                }
-                let save_enabled = !in_cell_pick || cell_has_sel;
-                if ui
-                    .add_enabled(save_enabled, egui::Button::new("Save"))
-                    .clicked()
-                {
-                    save = true;
-                }
-            });
-        });
-
-    if !open || cancel {
-        *picker = ActivePicker::None;
-        return PickerResult::None;
-    }
-    if back {
-        if let Some(slot) = picker.cell_pick_mut() {
-            *slot = None;
-        }
-        return PickerResult::None;
-    }
-    if save {
-        if in_cell_pick {
-            // Collection cell Save commits immediately.
-            let staged = picker
-                .cell_pick_mut()
-                .and_then(|c| c.as_ref())
-                .and_then(|p| p.to_ref());
-            if let Some(coord) = staged {
-                result = match picker.coord_kind() {
-                    Some(CoordKind::Point) => PickerResult::Point(coord),
-                    Some(CoordKind::SearchArea) => PickerResult::SearchArea(coord),
-                    None => PickerResult::None,
-                };
-                *picker = ActivePicker::None;
-            }
-            return result;
-        }
-        result = match picker {
-            ActivePicker::Items { staged, .. } => PickerResult::Items(staged.clone()),
-            ActivePicker::Coord {
-                kind: CoordKind::Point,
-                value,
-                ..
-            } => PickerResult::Point(CoordinateRef(value.clone())),
-            ActivePicker::Coord {
-                kind: CoordKind::SearchArea,
-                value,
-                ..
-            } => PickerResult::SearchArea(CoordinateRef(value.clone())),
-            ActivePicker::Macro { value, .. } => PickerResult::MacroName(value.clone()),
-            ActivePicker::Window {
-                process_path,
-                window_title,
-                ..
-            } => PickerResult::Window {
-                process_path: process_path.clone(),
-                window_title: window_title.clone(),
-            },
-            ActivePicker::None => PickerResult::None,
-        };
-        *picker = ActivePicker::None;
-    }
-    result
-}
-
-#[derive(Debug, Clone)]
-pub enum PickerResult {
-    None,
-    Items(Vec<String>),
-    Point(CoordinateRef),
-    SearchArea(CoordinateRef),
-    MacroName(String),
-    Window {
-        process_path: String,
-        window_title: String,
-    },
-}
-
-/// Static option lists for ComboBox fields (>2 options).
-pub mod options {
-    pub const LOOP_JUMP_MODES: &[(&str, &str)] = &[("break", "Break"), ("continue", "Continue")];
-
-    /// Match-order options as `(stored value, display label)`.
-    pub const ORDER_GROUPING: &[(&str, &str)] =
-        &[("row", "Row"), ("column", "Column"), ("none", "None")];
-    pub const ORDER_HORIZONTAL: &[(&str, &str)] = &[
-        ("left_to_right", "Left → Right"),
-        ("right_to_left", "Right → Left"),
-    ];
-    pub const ORDER_VERTICAL: &[(&str, &str)] = &[
-        ("top_to_bottom", "Top → Bottom"),
-        ("bottom_to_top", "Bottom → Top"),
-    ];
-
-    pub const SELECT_DEVICES: &[&str] = &["", "mouse", "keyboard"];
-    pub const SELECT_PRESS_MODES: &[&str] = &["", "click", "down", "up", "hold"];
-    pub const MOUSE_BUTTONS: &[&str] = &["", "left", "right", "center"];
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        item_tooltip_parts, query_matches_name_or_tags, query_matches_window, sort_by_display_name,
-        toggle_select_all_filtered,
-    };
-    use sqyre_capture::WindowInfo;
-    use sqyre_persist::{ProgramCatalog, ProgramData, ProgramItem};
-    use std::collections::BTreeMap;
-
-    #[test]
-    fn sort_by_display_name_orders_case_insensitive() {
-        let mut rows = vec![
-            ("b".into(), "Zebra".into()),
-            ("a".into(), "apple".into()),
-            ("c".into(), "Banana".into()),
-        ];
-        sort_by_display_name(&mut rows);
-        assert_eq!(
-            rows.iter().map(|(_, d)| d.as_str()).collect::<Vec<_>>(),
-            vec!["apple", "Banana", "Zebra"]
-        );
-    }
-
-    #[test]
-    fn toggle_select_all_adds_then_clears_filtered() {
-        let filtered = vec!["A~1".into(), "A~2".into()];
-        let mut selected = vec!["B~9".into()];
-        toggle_select_all_filtered(&mut selected, &filtered);
-        assert!(selected.contains(&"A~1".into()));
-        assert!(selected.contains(&"A~2".into()));
-        assert!(selected.contains(&"B~9".into()));
-        toggle_select_all_filtered(&mut selected, &filtered);
-        assert_eq!(selected, vec!["B~9".to_string()]);
-    }
-
-    #[test]
-    fn empty_query_matches_anything() {
-        assert!(query_matches_name_or_tags("", "Potion", &[]));
-        assert!(query_matches_name_or_tags("", "x", &["healing".into()]));
-    }
-
-    #[test]
-    fn matches_name_substring() {
-        assert!(query_matches_name_or_tags("pot", "HealthPotion", &[]));
-        assert!(!query_matches_name_or_tags("sword", "HealthPotion", &[]));
-    }
-
-    #[test]
-    fn matches_name_fuzzy_subsequence() {
-        assert!(query_matches_name_or_tags("hlt", "HealthPotion", &[]));
-        assert!(query_matches_name_or_tags("HPT", "HealthPotion", &[]));
-        assert!(!query_matches_name_or_tags("thl", "HealthPotion", &[])); // wrong order
-    }
-
-    #[test]
-    fn matches_tag_substring() {
-        let tags = vec!["consumable".into(), "healing".into()];
-        assert!(query_matches_name_or_tags("heal", "Minor Flask", &tags));
-        assert!(query_matches_name_or_tags("CONSUM", "Minor Flask", &tags));
-        assert!(!query_matches_name_or_tags("weapon", "Minor Flask", &tags));
-    }
-
-    #[test]
-    fn item_tooltip_parts_resolves_name_and_tags() {
-        let mut cat = ProgramCatalog::default();
-        cat.programs_mut().insert(
-            "Game".into(),
-            ProgramData {
-                name: "Game".into(),
-                items: BTreeMap::from([(
-                    "Flask".into(),
-                    ProgramItem {
-                        name: "Health Flask".into(),
-                        tags: vec!["healing".into(), "consumable".into()],
-                        ..Default::default()
-                    },
-                )]),
-                ..Default::default()
-            },
-        );
-
-        let (name, tags) = item_tooltip_parts(&cat, "Game~Flask");
-        assert_eq!(name, "Health Flask");
-        assert_eq!(tags, vec!["healing", "consumable"]);
-
-        let (name, tags) = item_tooltip_parts(&cat, "Game~Flask~v2");
-        assert_eq!(name, "Health Flask");
-        assert_eq!(tags, vec!["healing", "consumable"]);
-
-        let (name, tags) = item_tooltip_parts(&cat, "Missing~Item");
-        assert_eq!(name, "Item");
-        assert!(tags.is_empty());
-    }
-
-    #[test]
-    fn window_query_matches_title_name_or_path() {
-        let w = WindowInfo {
-            title: "Inbox — Mail".into(),
-            process_name: "thunderbird".into(),
-            process_path: "/usr/lib/thunderbird/thunderbird".into(),
-            icon: None,
-        };
-        assert!(query_matches_window("", &w));
-        assert!(query_matches_window("inbox", &w));
-        assert!(query_matches_window("THUNDER", &w));
-        assert!(query_matches_window("/usr/lib/thunder", &w));
-        assert!(!query_matches_window("firefox", &w));
-    }
-}
