+++ crates/sqyre-ui-model/src/lib.rs
@@ -5,6 +5,7 @@
 //! adds presentation (pill/tree/color) on top.
 
 mod action_log;
+mod action_picker;
 mod colors;
 mod display;
 mod highlight;
@@ -12,6 +13,7 @@ mod icons;
 mod runtime_vars;
 
 pub use action_log::*;
+pub use action_picker::*;
 pub use colors::*;
 pub use display::*;
 pub use highlight::*;
